Basic testing for reading workplace contents.
Remade package to distinguish functionality - specifically removed the
developermanagement (not developermanagement.handlers) project and added
the projectanalysis project.
Created and added 'ProjectStructureTesterHandler' handler for
projectanalysis to the plugin. As noted in the comments for this
handler, if anyone else runs into problems testing their code that
requires a UI context (my termonology is probably wrong here), feel free
to repurpose this handler for their own testing usage.

Very basic reading of the contents of the projects in the workplace is
now implemented. I had particular trouble with this as whenever I tested
my code in a standard psvMain method, it would be impossible for me to
obtain a reference to the Workplace or Workbench through the standard
API. I eventually realized, after looking through all the examples I
could find online, that I needed to use the API to access the workplace
within the context of the plugin.